refactor: room 패키지 내부에 있던 회계 정보를 독립적인 패키지로 구성
