+++ src/main/java/com/travel/role/domain/accounting/entity/AccountingInfo.java
@@ -1,5 +1,7 @@
-package com.travel.role.domain.board.entity;
+package com.travel.role.domain.accounting.entity;
 
+import com.travel.role.domain.board.entity.Board;
+import com.travel.role.domain.board.entity.BookInfo;
 import com.travel.role.domain.room.entity.Room;
 import lombok.*;
 
@@ -48,5 +50,4 @@ public class AccountingInfo {
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "book_info_id")
     private BookInfo bookInfo;
-
 }
