+++ src/main/java/com/travel/role/domain/accounting/repository/AccountingInfoRepository.java
@@ -1,6 +1,6 @@
-package com.travel.role.domain.board.repository;
+package com.travel.role.domain.accounting.repository;
 
-import com.travel.role.domain.board.entity.AccountingInfo;
+import com.travel.role.domain.accounting.entity.AccountingInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AccountingInfoRepository extends JpaRepository<AccountingInfo, Long> {
