+++ src/main/java/com/travel/role/domain/board/dto/request/BoardRequestDTO.java
@@ -1,7 +1,7 @@
 package com.travel.role.domain.board.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.travel.role.domain.board.entity.Category;
+import com.travel.role.domain.accounting.entity.Category;
 import com.travel.role.global.exception.dto.ExceptionMessage;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
